+++ src/UglyToad.PdfPig/Writer/PdfDedupStreamWriter.cs
@@ -13,7 +13,7 @@
             Stream stream,
             bool dispose,
             ITokenWriter tokenWriter = null,
-            Action<decimal> recordVersion = null
+            Action<double> recordVersion = null
             ) : base(stream, dispose, tokenWriter, recordVersion)
         {
         }
